autonomoose outline
this should be a general outline of goals for having the husky lens (which has its own program and will input values of ID and distance on its own into this code) input its data into our code. I also want to set this up with being able to detect which of 3 positions the object is in and associating it with a memorized apriltag (this apriltag isn't setup but should be the easier latter part) message me for any questions and further goal discussion.